+++ application/views/fleet/page.php
@@ -1,3 +1,0 @@
-{header}
-    {plane_list}
-{footer}