+++ CHANGELOG.md
@@ -13,10 +13,25 @@ All notable changes to this project will be documented in this file.
 
 # [latest stable version 1.0.0]
 
+## [2.0.0] - 2017-11-11
+## Changed
+
+## Added
+### Models
+ - Entity Model for base class of entities.
+ - Flight.php use as base class for all flight instances.
+ - Plane.php use as base class for all plane instances.
+
+## [2.0.0] - 2017-11-10
+## Changed
+ - Rearrange fleet and plane view files.
+
 ## [2.0.0] - 2017-11-09
 ## Changed
  - Add user role selection to the menu bar.
  - Display user role on page title and an indicator on the role drop-down.
+ - Modify page views for data mapping from CSV.
+ - Modify flight and fleet models to extend from an appropriate CSV model.
 ## Added
 ### Controllers
  - User role controller.